The inverted index can get big for 10s of thousands of larger files, but a near duplicate scan can still complete in under a few minutes on some server class machine from 2015 using only about 16 GiB of RAM.  This particular invidx module uses ~16 bytes per unique digest for a digest key and an inline 4-vector of 2-byte file numbers. (The excess over 16B depends on the collision/near duplication statistics, but it is tiny unless you have such rampant near-duplication that this tool is unlikely to help in the first place.)
In comparison the Nim `std/tables.Tables[Digest,seq[]]` have

  - a hash code (that, yes, I added myself) which is unhelpful for fast-
    to-hash integer keys and no way to disable that like my LPTabz

  - The seq is a pointer and a length adding another 8B + 8B

  - Allocation overhead is non-trivial where more slots are allocated
    for a seq than strictly needed.

instead of 16B/unique digest the space is more like 24+alloc extra
overhead.  With `-d:useMalloc`, for example, there is usually an 8B
header specifying the length of the region for efficient `realloc`.
So, in that mode you are looking at (32+16)/16 =~ 3X the space.  A
measurement I did showed more like 3.6X the space with the residual
likely related to "extra space" for each allocation area itself.
This is probably close to a best case with a worst case being more
like 6X the space.

Since memory for this is the dominant RAM cost here and swapping is
extremely hostile to random/hash-oriented workloads, this specialized
module expands the problems you can reach on a single machine by about
that same factor of 3-6X.