`ffprobe` is not very fast (maybe 50 ms/invocation) making final UNRESOLVED REPORT phase slow.  Instead cache `bppd` output in a new `$ndv/met` hierarchy like the `set.NL`/ one.
`bppd` is a Bits Per-Pixel,Date space factorizer.  This also removes
dependence upon the `lc` & `nio` packages from the default config.
We update `vhup` to just use this and drop a lot from `vdup`.

Along for the ride:

 - simple dumping of met/$path to stdout makes atomic pipe writes
   easy, though things are so fast now it really doesn't buy much.

 - `grep -vf "$kd"`, the preprocessed filter much earlier

 - `sort -nr` (`-r` being salient) before saving in `ISect` allowing
   the new all-in-one `awk` script in the merged `siMeta` to exit at
   the first match which is much less processing than prior version
   last in sort -n order.

 - the `awk` also just does 2 strcmps instead of yet another `grep`.

Altogether, this made this phase of a test case over 30x faster.